Wrote AST parser.
So I initially thought I'd write my interpreter and code generator to
operate on lisp Literals directly, but that proved to be pretty
difficult with all the because evaluation code and special form error
checking would be right next to each other, producing some pretty bulky
code. The code is pretty bulky to begin with, but hopefully it'll lead
to less bulky code later.