Implemented simple interpreter.
Currently doesn't handle functions. This is because the question of
functions is complicated by the builtin functions that the VM uses. The
interpreter and the VM pass arguments in very different ways, but both
use environments. The compiler will care about arity, but the VM
doesn't. The VM is ultimately the important part, so I guess everybody
else will have to adjust.