Fixed old bug with str_to_ast and vectors.
The parser parses into a vector of literals. The ast parser parses single
literals to single ASTs. `str_to_vec` mapped the ast parser over the literals,
and returned a vector. A vector isn't a real AST, so every consumer of
`str_to_ast` had to either be mapped over the vector, or explicitly wrap it in
an `AST::Do` to make it into a single AST value again. So I wrote `parse_multi`
which does the mapping and wrapping automatically.