Wrote `passes::list` to deal with `list` calls.
Variadic arguments aren't currently supported, and most of my "proposals" on how
to handle them depend on a way to create lists at runtime, and all variadicy has
to come from somewhere, so I implemented `list` as an AST pass to convert to
nested `cons` calls. Also incorporated it into the repl and the inspector, which
reports whether the pass made changes, but not where or how many, because I'm
kind of lazy. An AST differ might be useful, but an AST -> Literal converter
would be more useful, and a Literal differ would also be useful.