Added print syscall, and fixed related bug
I realized I needed _some_ way of testing multiple things at the same
time, so I added the print syscall. However, I ran into a problem: when
running things in the command line execution environment, the first call
to print always seemed to print twice. As it turns out, I had a bug in
the entrypoint repl/run interaction with `VM::import_jump`.
`import_jump` pushed the address of the new code to the stack. When a
`VM` is created, it starts with `(0, 0)` already on its stack. When the
first code chunk from the repl gets loaded into the VM, and then
returns, it pops the address of the chunk, and keeps running from the
initial `(0, 0)`, effectively executing the first chunk twice. This bug
has been in place since I introduced `import_jump`, I just never noticed
due to the lack of side effects. Hmmm.

To solve this issue, I had to break out gdb, because I didn't think to
look up how to print arbitrary backtraces in Rust without panicing,
because it's almost 3am.

My first guess was that I had a logic error that was causing double
syscall invocations, so set a breakpoint for the print function.
Incidentally, this confirmed it was actually a double syscall
invocation, and not something else printing (like an errant debug
statement). I could hit the breakpoint twice in the first chunk, and
only once in subsequent chunks, which narrowed the bug further. So I
looked at the backtrace:

```

```

They're identical. So exactly the same executiong path is invoking the
syscal, so the instructions being executed in the VM are somehow double
invoking the syscall? No, the dissassembly for this chunk of code was
entirely mundane: 1 chunk, no jumps, only 1 call instruction, no weird
stack interactions.

So worked my way up my stack to check the state of the VM. I set up 2
separate gdb instances, and moved them forward to the 2 separate
invocations. Using the gdb command:

```
print (*self).frames
```

I investigated the state of the VM's frame stack, and found that they
disagreed on the number of frames. Using the command:

```
print *((*self).frames.buf.ptr.pointer.0 + 0)
```

I discovered that the first frame stack had an extra `(0, 0)` at the
bottom of the stack that the second invocation didn't have. But above
the `(0, 0)`, the stacks where identical. This is when I remembered that
VMs are created with `(0, 0)` already on the stack (this is an artifact
from a time when VMs weren't reusable; you can also see this in the VM
documentation right now, because I haven't updated it).

How do I solve this? VM's could simply not spawn with anything on the
frame stack, but then you'd have to put it there manually. I could force
users to install valid bytecode in the VM, then run the VM forward until
the dummy code returned, clearing the stack. However, I decided that the
`import_jump` function should clear the frames when called.

The fact that it doesn't is inherently weird, and there just isn't a
usecase for pausing the VM in the middle of a function, forcing it to
call another function, and then continuing execution. You'd somehow have
to handle the extra value(s) on the data stack. Maybe I was concerned
about "higher order syscalls", like if I wanted to implement `map` as a
syscall, I might want a way to invoke arbitrary functions. But it's
unclear how the syscall was supposed to regain control from VM. If the
call stack looks like

```
vm -> syscall -> vm::invoke_function -> function returns
```

When the function returns, the syscall will still be on the stack, so
the syscall will be reentered fresh, but without any arguments on the
stack anymore. Higher order syscalls are an open problem right now, and
might not be possible. That sounds sort of crazy: how is my host
language supposed to interact with a client language? Currently, you can
invoke functions from Rust in the VM, as long as you don't do it in the
middle of another function. However, it will be possible to interact
with the messaging system fully, and forcing the host language to
architect its client communcation in that way would be annoying, but
principaled.

Anyway. `VM::import_jump` now clears the frame stack. Woe betide the
data stack though. There isn't even a way to pop it from the repl. It
just builds up.

The VM should probably clear both whenever it encounters an error, just
to prevent stupid errors, but that's in the future.