Worked on self hosted interpreter, made changes to suppor that.
I moved the interpreter to an env-threaded pattern to ensure env changes are
propagated. Not all defs are local.

I also discovered that `clojure-mode` in emacs is the best way to format ISL
code, but that it's very particular in its indenting. If `let` bindings aren't
in brackets, it doesn't indent correctly, so I added brackets to the list of
acceptable delimiters. Also, `lambda` doesn't exist in Clojure, it uses `fn`,
and `clojure-mode` formats `fn` differently from other function applications, so
I aliased `fn` in the _AST_ parser. Nowhere else though.