Wrote a test to run the self hosted interpreter through itself.
You know, just to see what would happen. It asserts false, so it
autofails, but it still wouldn't work even if I removed it. To solve
this issue, I've been writing dynamic syscall dispatch in the
interpreter, because I refuse to add actual dynamic dispatch. I will not
taint the VM with that shit.