Added DefVisitor trait.
I noticed this pattern a while ago: because several AST elements included
binding requests, and the request format was identical, I moved to an external
struct. I didn't incorporate defs into the ASTVisitor, because it was an
external struct, so I ended up writing def visitor methods in a lot of
ASTVisitors to make it easier. So I wrote a DefVisitor trait, and added a helper
method for visiting multiple defs as well as single defs. Then I refactored a
bunch of ASTVisitors to implement and use DefVisitor.