Reworked procinfo to be dynamic dispatch, added spawning.
ProcInfo, the mechanism by which VMs interact with their execution
environment, used to be a hardcoded struct. It is now a trait object,
permitting some dynamism and polymorphism. This also made it a lot
easier to separate concerns between the futures based executor and the
VM. They're still pretty tightly bound, but it's mostly 1 way, executor
depending on broad and deep access to the VM, while the VM relies only
on the very small ExecHandle trait. ExecHandle has the standard get_pid
and send message methods, but also I also added and implemented a
"spawn" method. It currently spawns an arbitrary VM on the same executor
(using tokio::spawn). The VM doesn't use this method, and the VM
derivation will be a little odd.